requirements.txt
fastapi>=0.117.1
uvicorn[standard]==0.30.6
openai>=1.30.0
mammoth>=1.6.0
python-multipart>=0.0.9
python-dotenv
